Move `blocks` to `assets/`.
