+++ .github/workflows/e2e-tests.yml
@@ -14,7 +14,6 @@ on:
       - '.github/workflows/e2e-tests.yml'
       - '.babelrc'
       - 'assets/**'
-      - 'blocks/**'
       - 'includes/**'
       - '/*.php'
       - 'jest-puppeteer.config.js'
@@ -47,7 +46,6 @@ on:
       - '.github/workflows/e2e-tests.yml'
       - '.babelrc'
       - 'assets/**'
-      - 'blocks/**'
       - 'includes/**'
       - '/*.php'
       - 'jest-puppeteer.config.js'
