+++ .github/workflows/storybook.yml
@@ -12,7 +12,6 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - 'blocks/**'
       - 'storybook/**'
       - '**.scss'
       - '.nvmrc'
@@ -30,7 +29,6 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - 'blocks/**'
       - 'storybook/**'
       - '**.scss'
       - '.nvmrc'
