+++ .github/workflows/zips.yml
@@ -15,7 +15,6 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - 'blocks/**'
       - 'includes/**'
       - 'google-site-kit.php'
       - 'scoper.inc.php'
@@ -42,7 +41,6 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - 'blocks/**'
       - 'includes/**'
       - 'google-site-kit.php'
       - 'scoper.inc.php'
