+++ assets/blocks/reader-revenue-manager/block-editor-plugin/editor-styles.scss
