+++ assets/blocks/reader-revenue-manager/block-editor-plugin/index.js
