+++ assets/blocks/reader-revenue-manager/common/constants.js
