+++ assets/blocks/reader-revenue-manager/common/editor-styles.scss
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-@import "../../../assets/sass/config/variables";
+@import "../../../../assets/sass/config/variables";
 
 // Styles copied over from the actual inline button.
 .googlesitekit-blocks-reader-revenue-manager-button {
