+++ assets/blocks/reader-revenue-manager/common/index.js
