+++ assets/blocks/reader-revenue-manager/contribute-with-google/block.json
