+++ assets/blocks/reader-revenue-manager/subscribe-with-google/Edit.js
@@ -30,7 +30,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import { ButtonEdit } from '../common';
-import { MODULES_READER_REVENUE_MANAGER } from '../../../assets/js/modules/reader-revenue-manager/datastore/constants';
+import { MODULES_READER_REVENUE_MANAGER } from '../../../../assets/js/modules/reader-revenue-manager/datastore/constants';
 
 /**
  * Subscribe with Google Block Edit component.
