+++ assets/blocks/reader-revenue-manager/subscribe-with-google/block.json
