+++ assets/blocks/sign-in-with-google/block.json
