+++ webpack/gutenbergBlocks.config.js
@@ -43,20 +43,21 @@ module.exports = ( mode ) => ( {
 	entry: {
 		// Reader Revenue Manager
 		'reader-revenue-manager/block-editor-plugin/index':
-			'./blocks/reader-revenue-manager/block-editor-plugin/index.js',
+			'./assets/blocks/reader-revenue-manager/block-editor-plugin/index.js',
 		'reader-revenue-manager/block-editor-plugin/editor-styles':
-			'./blocks/reader-revenue-manager/block-editor-plugin/editor-styles.scss',
+			'./assets/blocks/reader-revenue-manager/block-editor-plugin/editor-styles.scss',
 		// Reader Revenue Manager blocks.
 		'reader-revenue-manager/contribute-with-google/index':
-			'./blocks/reader-revenue-manager/contribute-with-google/index.js',
+			'./assets/blocks/reader-revenue-manager/contribute-with-google/index.js',
 		'reader-revenue-manager/subscribe-with-google/index':
-			'./blocks/reader-revenue-manager/subscribe-with-google/index.js',
+			'./assets/blocks/reader-revenue-manager/subscribe-with-google/index.js',
 		'reader-revenue-manager/common/editor-styles':
-			'./blocks/reader-revenue-manager/common/editor-styles.scss',
+			'./assets/blocks/reader-revenue-manager/common/editor-styles.scss',
 		// Sign in with Google block.
-		'sign-in-with-google/index': './blocks/sign-in-with-google/index.js',
+		'sign-in-with-google/index':
+			'./assets/blocks/sign-in-with-google/index.js',
 		'sign-in-with-google/editor-styles':
-			'./blocks/sign-in-with-google/editor-styles.scss',
+			'./assets/blocks/sign-in-with-google/editor-styles.scss',
 	},
 	externals: gutenbergExternals,
 	output: {
@@ -100,10 +101,10 @@ module.exports = ( mode ) => ( {
 		new CopyWebpackPlugin( {
 			patterns: [
 				{
-					from: 'blocks/**/block.json',
+					from: 'assets/blocks/**/block.json',
 					to: ( { context, absoluteFilename } ) => {
 						return absoluteFilename.replace(
-							`${ context }/blocks`,
+							`${ context }/assets/blocks`,
 							`${ rootDir }/dist/assets/js/blocks`
 						);
 					},
